refactored survey duration tracking helper methods.
